requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
fastapi-mail
python-multipart
slowapi
python-jose
passlib
fastapi-utils
